requirements.txt
slack_bolt==1.19.0
slack_sdk==3.33.2
python-dotenv==1.0.1
pandas==2.2.2
openai==1.101.0 
uvloop==0.19.0
groq==0.11.0
tenacity==9.0.0